+++ domain/entities/bus_sub_route.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: false
+
+require 'dry-struct'
+require_relative '../motc_mappers/bus_sub_route_mapper.rb'
+
+module TaiGo
+  module Entity
+    # Domain entity object for git BusRoute
+    class BusSubRoute < Dry::Struct
+      attribute :sub_route_uid, Types::Strict::String
+      attribute :route_id, Types::Strict::String
+      attribute :sub_route_name, Types.Instance(TaiGo::MOTC::BusSubRouteMapper::DataMapper::Name)
+      attribute :headsign, Types::String.optional
+      attribute :direction, Types::Int.optional
+    end
+  end
+end
