+++ domain/init.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-folders = %w[database_repositories]
+folders = %w[entities database_repositories motc_mappers]
 folders.each do |folder|
   require_relative "#{folder}/init.rb"
 end
