+++ infrastructure/init.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: false
 
-folders = %w[database/orm]
+folders = %w[motc database/orm]
 folders.each do |folder|
   require_relative "#{folder}/init.rb"
 end
