+++ infrastructure/motc/init.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: false
+
+require_relative 'motc_api.rb'
