+++ init.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-folders = %w[config entities domain infrastructure lib]
+folders = %w[config infrastructure domain]
 folders.each do |folder|
   require_relative "#{folder}/init.rb"
 end
