--- lib/init.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: false
-
-require_relative 'motc_api.rb'
-require_relative 'mapper/init.rb'
