--- lib/mapper/init.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: false
-
-Dir.glob("#{File.dirname(__FILE__)}/*.rb").each do |file|
-  require file
-end
